feat: add SplitPay contract
